backend/requirements.txt
boto3
botocore
Django
django-cors-headers
django-storages
dotenv
numpy
opencv-python
pandas
pillow
psycopg2-binary
PyYAML
requests
scikit-image
pydicom
